# routing, channeldb: several optimizations for path finding