@joostjager 2019-08-26
Good changes again :+1: 

Main question is what to do with the added `if node == nil` and `if db == nil` code. I think that with the right design this should be avoidable, but it probably requires more changes. On the other hand, building on top of something that shouldn't be there in the first place defers paying those costs to the future.

Final comment is about commit structure. In the end, these commits should be reorganized so that it becomes a logical stack. For example (some of these are already a separate commit now):
* all the cache changes in `graph.go` can be done first in one or multiple commits first. 
* add `NumNodes` to `graph.go` and pre-allocate map
* merging nextNode into the path finding heap
* change to nodeWithDist pointers
* remove unused k-shortest code
* remove distance map infinity initialization
* fix log performance

What you generally want to avoid in one pr is to make changes in a commit and change them again later. For example, initializing the map first with 10000 and later replacing it by `NumNodes`. I find it easier if those two are just squashed together and there is one commit that presents the final changes in a particular area.

You could also consider making two prs. One with the internal path finding improvements that are pretty straight-forward to review and get approval on. Then a second one to carefully do the cache.