@halseth 2019-10-09
@champo This is close! Do you want to make the last fixups? :)