@champo 2019-10-09
@halseth I'll work on it this weekend :D