@champo 2019-10-23
@halseth @joostjager updated with the estimated node count changes and linter fixes :)