@champo 2019-10-25
@halseth nits taken care of!

@joostjager I've rebased against master and fixed that error. It seems the commit that removed the old heap struct was already in master too and that brought in a silent conflict